requirements.txt
albumentations==1.3.1
appdirs==1.4.4
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.1
filelock==3.12.2
fsspec==2023.6.0
gitdb==4.0.10
gitpython==3.1.31
huggingface-hub==0.15.1
idna==3.4
imageio==2.31.1
joblib==1.2.0
lazy-loader==0.2
munch==3.0.0
networkx==3.1
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==23.1
pathtools==0.1.2
pretrainedmodels==0.7.4
protobuf==4.23.3
psutil==5.9.5
pywavelets==1.4.1
pyyaml==6.0
qudida==0.0.4
requests==2.31.0
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
segmentation-models-pytorch==0.3.3
sentry-sdk==1.25.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.9.2
tqdm==4.65.0
urllib3==2.0.3
wandb==0.15.4
yacs==0.1.8
